split server.py
