+++ src/carapace/server/__init__.py
@@ -5,8 +5,7 @@ import contextlib
 import logging  # stdlib logging used only for _InterceptHandler → loguru bridge
 import os
 import sys
-import uuid
-from collections.abc import AsyncIterator
+from collections.abc import AsyncGenerator
 from contextlib import asynccontextmanager
 from datetime import UTC, datetime, timedelta
 from decimal import Decimal
@@ -25,83 +24,53 @@ from fastapi import (
     Query,
     Request,
     Response,
-    WebSocket,
-    WebSocketDisconnect,
-    WebSocketException,
-    status,
 )
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
 from genai_prices import UpdatePrices
 from loguru import logger
 from pydantic import BaseModel, ValidationError, field_validator, model_validator
 from pydantic_ai.exceptions import UsageLimitExceeded
 from pydantic_ai.messages import ModelMessage
 
-from . import get_version
-from .auth import get_token
-from .bootstrap import ensure_data_dir, ensure_knowledge_dir
-from .cache import SessionListCache
-from .config import _resolve_data_dir, _resolve_knowledge_dir, get_config_path, get_data_dir, load_config
-from .credentials import CredentialRegistry, build_credential_registry
-from .git.http import GitHttpHandler
-from .git.store import GitStore
-from .jobs import JobsScheduler, JobsStore, build_job_run_message
-from .llm import make_model_factory
-from .models.config import Config
-from .models.jobs import JobDefinition, JobsFile
-from .models.session import SessionAttributes, SessionJobRunContext, SessionState
-from .models.tooling import ToolResult, normalize_tool_call_args
-from .notifications.models import (
-    NotificationClientType,
-    NotificationFocusState,
-    NotificationPreferences,
-    NotificationSubscription,
-)
-from .notifications.presence import NotificationPresenceRegistry
-from .notifications.router import NotificationRouter
-from .notifications.sender import WebPushSender
-from .notifications.store import NotificationStore, derive_owner_key
-from .notifications.vapid import derive_vapid_public_key, ensure_vapid_config
-from .sandbox.manager import SandboxManager
-from .sandbox.proxy import ProxyServer
-from .sandbox.runtime import ContainerRuntime
-from .sandbox.state import SessionSandboxSnapshot
-from .security.context import ApprovalSource, ApprovalVerdict
-from .session import SessionEngine, SessionManager
-from .session.archive import SessionArchiveService
-from .skills import SkillRegistry
-from .usage import LlmRequestState, SessionBudgetExceededError
-from .ws_models import (
-    SLASH_COMMANDS,
-    ApprovalRequest,
-    ApprovalResponse,
-    Cancelled,
-    CancelRequest,
-    CommandResult,
-    CredentialApprovalRequest,
-    DomainAccessApprovalRequest,
-    Done,
-    ErrorMessage,
-    EscalationResponse,
-    FinalStatus,
-    GitPushApprovalRequest,
-    LlmActivity,
-    LlmActivityUpdate,
-    ResetToTurnRequest,
-    RetryLatestTurnRequest,
-    ServerEnvelope,
-    SessionTitleUpdate,
-    StatusUpdate,
-    ThinkingChunk,
-    TokenChunk,
-    ToolCallInfo,
-    ToolResultInfo,
-    TurnUsage,
-    UserMessage,
-    UserMessageNotification,
-    parse_client_message,
-)
+from .. import get_version
+from ..auth import get_token
+from ..bootstrap import ensure_data_dir, ensure_knowledge_dir
+from ..cache import SessionListCache
+from ..config import _resolve_data_dir, _resolve_knowledge_dir, get_config_path, get_data_dir, load_config
+from ..credentials import CredentialRegistry, build_credential_registry
+from ..git.http import GitHttpHandler
+from ..git.store import GitStore
+from ..jobs import JobsScheduler, JobsStore, build_job_run_message
+from ..llm import make_model_factory
+from ..models.config import Config
+from ..models.jobs import JobDefinition, JobsFile
+from ..models.session import SessionAttributes, SessionJobRunContext, SessionState
+from ..models.tooling import normalize_tool_call_args
+from ..notifications.presence import NotificationPresenceRegistry
+from ..notifications.router import NotificationRouter
+from ..notifications.sender import WebPushSender
+from ..notifications.store import NotificationStore, derive_owner_key
+from ..notifications.vapid import ensure_vapid_config
+from ..sandbox.manager import SandboxManager
+from ..sandbox.proxy import ProxyServer
+from ..sandbox.runtime import ContainerRuntime
+from ..sandbox.state import SessionSandboxSnapshot
+from ..security.context import ApprovalSource, ApprovalVerdict
+from ..session import SessionEngine, SessionManager
+from ..session.archive import SessionArchiveService
+from ..skills import SkillRegistry
+from ..usage import SessionBudgetExceededError
+from ..ws_models import FinalStatus
+from .auth import verify_token as _verify_token
+from .auth import verify_ws_token
+from .notifications import _set_notification_presence as _set_notification_presence
+from .notifications import router as notifications_router
+from .websocket import ServerMeta as ServerMeta
+from .websocket import VapidPublicKeyResponse as VapidPublicKeyResponse
+from .websocket import WebSocketSubscriber as WebSocketSubscriber
+from .websocket import _llm_activity_payload as _llm_activity_payload
+from .websocket import _send as _send
+from .websocket import router as websocket_router
 
 load_dotenv()
 
@@ -123,12 +92,13 @@ _notification_router: NotificationRouter
 _SESSION_COMMIT_SWEEP_SECONDS = 15 * 60
 _JOB_SCHEDULER_SWEEP_SECONDS = 60
 _APP_VERSION = get_version()
+_verify_ws_token = verify_ws_token
 
 
 def _create_sandbox_runtime(config: Config, data_dir: Path) -> ContainerRuntime:
     """Instantiate the sandbox container runtime based on config."""
     if config.sandbox.runtime == "kubernetes":
-        from .sandbox.kubernetes import KubernetesRuntime
+        from ..sandbox.kubernetes import KubernetesRuntime
 
         return KubernetesRuntime(
             namespace=config.sandbox.k8s_namespace,
@@ -148,7 +118,7 @@ def _create_sandbox_runtime(config: Config, data_dir: Path) -> ContainerRuntime:
             resource_limits_memory=config.sandbox.k8s_resource_limits_memory,
         )
 
-    from .sandbox.docker import DockerRuntime
+    from ..sandbox.docker import DockerRuntime
 
     host_data_dir_env = os.environ.get("CARAPACE_HOST_DATA_DIR")
     return DockerRuntime(
@@ -237,7 +207,7 @@ async def _autosave_inactive_sessions() -> None:
 
 
 @asynccontextmanager
-async def _lifespan(app: FastAPI) -> AsyncIterator[None]:
+async def _lifespan(app: FastAPI) -> AsyncGenerator[None]:
     global \
         _data_dir, \
         _config, \
@@ -460,7 +430,7 @@ async def _lifespan(app: FastAPI) -> AsyncIterator[None]:
 
     matrix_channel = None
     if _config.channels.matrix.enabled:
-        from .channels.matrix import MatrixChannel
+        from ..channels.matrix import MatrixChannel
 
         matrix_channel = MatrixChannel(
             config=_config.channels.matrix,
@@ -516,250 +486,6 @@ app.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"],
 )
-
-_bearer_scheme = HTTPBearer()
-
-
-async def _verify_token(
-    credentials: Annotated[HTTPAuthorizationCredentials, Depends(_bearer_scheme)],
-) -> str:
-    expected = get_token()
-    if credentials.credentials != expected:
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid token")
-    return credentials.credentials
-
-
-async def _verify_ws_token(
-    websocket: WebSocket,
-    token: Annotated[str | None, Query()] = None,
-) -> str:
-    expected = get_token()
-    if token and token == expected:
-        return token
-    auth = websocket.headers.get("authorization", "")
-    if auth.startswith("Bearer ") and auth.removeprefix("Bearer ") == expected:
-        return expected
-    await websocket.close(code=status.WS_1008_POLICY_VIOLATION)
-    raise WebSocketException(code=status.WS_1008_POLICY_VIOLATION)
-
-
-# --- REST: Notifications ---
-
-
-class NotificationPreferencesPatch(BaseModel):
-    escalation_pending: bool | None = None
-    attended_turn_completed: bool | None = None
-    unattended_turn_completed: bool | None = None
-    unattended_turn_failed: bool | None = None
-
-    def apply(self, prefs: NotificationPreferences) -> NotificationPreferences:
-        return prefs.model_copy(update=self.model_dump(exclude_none=True))
-
-
-class NotificationSubscriptionCreateRequest(BaseModel):
-    endpoint: str
-    p256dh: str
-    auth: str
-    device_name: str = ""
-    preferences: NotificationPreferencesPatch | None = None
-
-    @field_validator("endpoint", "p256dh", "auth", "device_name", mode="before")
-    @classmethod
-    def _normalize_string_field(cls, value: str) -> str:
-        return value.strip()
-
-
-class NotificationSubscriptionResponse(BaseModel):
-    subscription_id: str
-    device_name: str
-    endpoint: str
-    subscribed_at: str
-    expires_at: str
-    last_heartbeat: str | None = None
-    preferences: NotificationPreferences
-
-
-class NotificationTestResponse(BaseModel):
-    delivered: bool
-
-
-class NotificationPresenceRequest(BaseModel):
-    session_id: str
-    client_type: NotificationClientType
-    focus_state: NotificationFocusState
-
-    @field_validator("session_id", mode="before")
-    @classmethod
-    def _normalize_session_id(cls, value: str) -> str:
-        normalized = value.strip()
-        if not normalized:
-            raise ValueError("session_id must not be empty")
-        return normalized
-
-
-class InteractivePresenceRequest(NotificationPresenceRequest):
-    source_id: str
-
-    @field_validator("source_id", mode="before")
-    @classmethod
-    def _normalize_source_id(cls, value: str) -> str:
-        normalized = value.strip()
-        if not normalized:
-            raise ValueError("source_id must not be empty")
-        return normalized
-
-
-def _notification_ttl() -> timedelta:
-    return timedelta(days=_config.notifications.subscription_ttl_days)
-
-
-def _default_notification_preferences() -> NotificationPreferences:
-    return _config.notifications.default_preferences.model_copy(deep=True)
-
-
-def _notification_response(subscription: NotificationSubscription) -> NotificationSubscriptionResponse:
-    return NotificationSubscriptionResponse(
-        subscription_id=subscription.id,
-        device_name=subscription.device_name,
-        endpoint=subscription.endpoint,
-        subscribed_at=subscription.subscribed_at.isoformat(),
-        expires_at=subscription.expires_at.isoformat(),
-        last_heartbeat=subscription.last_heartbeat.isoformat() if subscription.last_heartbeat is not None else None,
-        preferences=subscription.notification_prefs.model_copy(deep=True),
-    )
-
-
-def _owned_notification_subscription(subscription_id: str, token: str) -> NotificationSubscription:
-    subscription = _notification_store.get_subscription(subscription_id)
-    if subscription is None or subscription.owner_key != derive_owner_key(token):
-        raise HTTPException(status_code=404, detail="Notification subscription not found")
-    return subscription
-
-
-async def _set_notification_presence(
-    *,
-    session_id: str,
-    source_id: str,
-    client_type: NotificationClientType,
-    focus_state: NotificationFocusState,
-    now: datetime | None = None,
-) -> None:
-    if focus_state == "inactive":
-        _notification_presence.remove_presence(session_id=session_id, source_id=source_id)
-        return
-    _notification_presence.update_presence(
-        session_id=session_id,
-        source_id=source_id,
-        client_type=client_type,
-        focus_state=focus_state,
-        now=now,
-    )
-    if _notification_presence.is_session_actively_handled(session_id, now=now):
-        await _engine.clear_pending_notifications(session_id)
-
-
-@router.get("/notifications/subscriptions", response_model=list[NotificationSubscriptionResponse])
-async def list_notification_subscriptions(
-    _token: str = Depends(_verify_token),
-) -> list[NotificationSubscriptionResponse]:
-    _notification_store.cleanup_expired()
-    owner_key = derive_owner_key(_token)
-    subscriptions = _notification_store.list_subscriptions(owner_key=owner_key)
-    return [_notification_response(subscription) for subscription in subscriptions]
-
-
-@router.post("/notifications/subscriptions", response_model=NotificationSubscriptionResponse)
-async def upsert_notification_subscription(
-    request: NotificationSubscriptionCreateRequest,
-    _token: str = Depends(_verify_token),
-) -> NotificationSubscriptionResponse:
-    _notification_store.cleanup_expired()
-    prefs = _default_notification_preferences()
-    if request.preferences is not None:
-        prefs = request.preferences.apply(prefs)
-    subscription = _notification_store.upsert_subscription(
-        owner_key=derive_owner_key(_token),
-        endpoint=request.endpoint,
-        p256dh=request.p256dh,
-        auth=request.auth,
-        device_name=request.device_name,
-        notification_prefs=prefs,
-        ttl=_notification_ttl(),
-    )
-    return _notification_response(subscription)
-
-
-@router.delete("/notifications/subscriptions/{subscription_id}", status_code=204)
-async def delete_notification_subscription(subscription_id: str, _token: str = Depends(_verify_token)) -> Response:
-    subscription = _owned_notification_subscription(subscription_id, _token)
-    _notification_store.delete_subscription(subscription.id)
-    return Response(status_code=204)
-
-
-@router.patch(
-    "/notifications/subscriptions/{subscription_id}/preferences",
-    response_model=NotificationSubscriptionResponse,
-)
-async def patch_notification_subscription_preferences(
-    subscription_id: str,
-    request: NotificationPreferencesPatch,
-    _token: str = Depends(_verify_token),
-) -> NotificationSubscriptionResponse:
-    subscription = _owned_notification_subscription(subscription_id, _token)
-    updated = subscription.model_copy(update={"notification_prefs": request.apply(subscription.notification_prefs)})
-    _notification_store.save_subscription(updated)
-    return _notification_response(updated)
-
-
-@router.post(
-    "/notifications/subscriptions/{subscription_id}/test",
-    response_model=NotificationTestResponse,
-)
-async def test_notification_subscription(
-    subscription_id: str,
-    _token: str = Depends(_verify_token),
-) -> NotificationTestResponse:
-    _notification_store.cleanup_expired()
-    subscription = _owned_notification_subscription(subscription_id, _token)
-    delivered = await _notification_router.dispatch_test(subscription=subscription)
-    if not delivered:
-        raise HTTPException(status_code=502, detail="Failed to deliver test notification")
-    return NotificationTestResponse(delivered=True)
-
-
-@router.post("/notifications/subscriptions/{subscription_id}/presence")
-async def update_notification_presence(
-    subscription_id: str,
-    request: NotificationPresenceRequest,
-    _token: str = Depends(_verify_token),
-) -> dict[str, bool]:
-    subscription = _owned_notification_subscription(subscription_id, _token)
-    now = datetime.now(tz=UTC)
-    updated = subscription.model_copy(update={"last_heartbeat": now, "expires_at": now + _notification_ttl()})
-    _notification_store.save_subscription(updated)
-    await _set_notification_presence(
-        session_id=request.session_id,
-        source_id=subscription_id,
-        client_type=request.client_type,
-        focus_state=request.focus_state,
-        now=now,
-    )
-    return {"heartbeat_received": True}
-
-
-@router.post("/notifications/presence")
-async def update_interactive_presence(
-    request: InteractivePresenceRequest,
-    _token: str = Depends(_verify_token),
-) -> dict[str, bool]:
-    await _set_notification_presence(
-        session_id=request.session_id,
-        source_id=request.source_id,
-        client_type=request.client_type,
-        focus_state=request.focus_state,
-    )
-    return {"heartbeat_received": True}
-
 
 # --- REST: Sessions ---
 
@@ -901,14 +627,6 @@ class JobRunResult(BaseModel):
     session_id: str
     created_new_session: bool
     session: SessionInfo
-
-
-class ServerMeta(BaseModel):
-    version: str
-
-
-class VapidPublicKeyResponse(BaseModel):
-    vapid_public_key: str
 
 
 async def _run_job_definition(
@@ -1575,425 +1293,6 @@ def _history_from_messages(session_id: str) -> list[HistoryMessage]:
     return result
 
 
-# --- WebSocket: Chat ---
-
-
-async def _send(ws: WebSocket, msg: ServerEnvelope) -> None:
-    await ws.send_json(msg.model_dump(mode="json"))
-
-
-def _llm_activity_payload(activity: LlmRequestState | None) -> LlmActivity | None:
-    if activity is None:
-        return None
-    return LlmActivity(
-        request_id=activity.request_id,
-        source=activity.source,
-        model=activity.model_name,
-        phase=activity.phase,
-        started_at=activity.started_at,
-        first_thinking_at=activity.first_thinking_at,
-        last_thinking_at=activity.last_thinking_at,
-        first_text_at=activity.first_text_at,
-    )
-
-
-@router.get("/commands")
-async def list_commands(_token: str = Depends(_verify_token)) -> list[dict[str, str]]:
-    return SLASH_COMMANDS
-
-
-@router.get("/meta", response_model=ServerMeta)
-async def get_meta(_token: str = Depends(_verify_token)) -> ServerMeta:
-    return ServerMeta(version=_APP_VERSION)
-
-
-@router.get("/config/vapid-public-key", response_model=VapidPublicKeyResponse)
-async def get_vapid_public_key() -> VapidPublicKeyResponse:
-    vapid_private_key = _config.notifications.vapid_private_key
-    assert vapid_private_key is not None
-    vapid_public_key = derive_vapid_public_key(vapid_private_key)
-    return VapidPublicKeyResponse(vapid_public_key=vapid_public_key)
-
-
-@router.get("/models")
-async def list_models(_token: str = Depends(_verify_token)) -> list[dict[str, Any]]:
-    return [e.model_dump(mode="json", by_alias=True) for e in _engine.available_model_entries]
-
-
-class WebSocketSubscriber:
-    """Thin adapter: forwards ``SessionEngine`` events to a WebSocket."""
-
-    def __init__(self, ws: WebSocket) -> None:
-        self._ws = ws
-
-    async def _safe_send(self, msg: ServerEnvelope) -> None:
-        try:
-            await _send(self._ws, msg)
-        except Exception as exc:
-            logger.warning(f"WebSocket send failed: {exc}")
-
-    async def on_user_message(self, content: str, *, from_self: bool) -> None:
-        await self._safe_send(UserMessageNotification(content=content))
-
-    async def on_tool_call(
-        self,
-        tool: str,
-        args: dict[str, Any],
-        detail: str,
-        approval_source: ApprovalSource | None = None,
-        approval_verdict: ApprovalVerdict | None = None,
-        approval_explanation: str | None = None,
-        tool_id: str | None = None,
-        parent_tool_id: str | None = None,
-    ) -> None:
-        normalized_args = normalize_tool_call_args(tool, args)
-        contexts_raw = normalized_args.get("contexts")
-        contexts = list(contexts_raw) if isinstance(contexts_raw, list) else []
-        await self._safe_send(
-            ToolCallInfo(
-                tool=tool,
-                args=normalized_args,
-                detail=detail,
-                contexts=contexts,
-                approval_source=approval_source,
-                approval_verdict=approval_verdict,
-                approval_explanation=approval_explanation,
-                tool_id=tool_id,
-                parent_tool_id=parent_tool_id,
-            )
-        )
-
-    async def on_tool_result(self, result: ToolResult) -> None:
-        await self._safe_send(
-            ToolResultInfo(
-                tool=result.tool,
-                result=result.output,
-                exit_code=result.exit_code,
-                tool_id=result.tool_id,
-            )
-        )
-
-    async def on_token(self, content: str) -> None:
-        await self._safe_send(TokenChunk(content=content))
-
-    async def on_thinking_token(self, content: str) -> None:
-        await self._safe_send(ThinkingChunk(content=content))
-
-    async def on_llm_activity(self, activity: LlmRequestState | None) -> None:
-        await self._safe_send(LlmActivityUpdate(activity=_llm_activity_payload(activity)))
-
-    async def on_done(
-        self,
-        content: str,
-        usage: TurnUsage,
-        *,
-        thinking: str | None = None,
-        final_status: FinalStatus | None = None,
-    ) -> None:
-        await self._safe_send(Done(content=content, thinking=thinking, usage=usage, final_status=final_status))
-
-    async def on_error(self, detail: str, *, turn_terminal: bool = False) -> None:
-        await self._safe_send(ErrorMessage(detail=detail, turn_terminal=turn_terminal))
-
-    async def on_cancelled(self) -> None:
-        await self._safe_send(Cancelled())
-
-    async def on_approval_request(self, req: ApprovalRequest) -> None:
-        await self._safe_send(req)
-
-    async def on_domain_access_approval_request(self, request_id: str, domain: str, command: str) -> None:
-        await self._safe_send(DomainAccessApprovalRequest(request_id=request_id, domain=domain, command=command))
-
-    async def on_git_push_approval_request(
-        self, request_id: str, ref: str, explanation: str, changed_files: list[str]
-    ) -> None:
-        await self._safe_send(
-            GitPushApprovalRequest(request_id=request_id, ref=ref, explanation=explanation, changed_files=changed_files)
-        )
-
-    async def on_title_update(self, title: str, usage: TurnUsage | None = None) -> None:
-        await self._safe_send(SessionTitleUpdate(title=title, usage=usage))
-
-    async def on_domain_info(
-        self,
-        domain: str,
-        detail: str,
-        approval_source: ApprovalSource | None = None,
-        approval_verdict: ApprovalVerdict | None = None,
-        approval_explanation: str | None = None,
-        tool_id: str | None = None,
-        parent_tool_id: str | None = None,
-    ) -> None:
-        await self._safe_send(
-            ToolCallInfo(
-                tool="proxy_domain",
-                args={"domain": domain},
-                detail=detail,
-                approval_source=approval_source,
-                approval_verdict=approval_verdict,
-                approval_explanation=approval_explanation,
-                tool_id=tool_id,
-                parent_tool_id=parent_tool_id,
-            )
-        )
-
-    async def on_git_push_info(
-        self,
-        ref: str,
-        decision: str,
-        detail: str,
-        approval_source: ApprovalSource | None = None,
-        approval_verdict: ApprovalVerdict | None = None,
-        approval_explanation: str | None = None,
-        tool_id: str | None = None,
-        parent_tool_id: str | None = None,
-    ) -> None:
-        await self._safe_send(
-            ToolCallInfo(
-                tool="git_push",
-                args={"ref": ref, "decision": decision},
-                detail=detail,
-                approval_source=approval_source,
-                approval_verdict=approval_verdict,
-                approval_explanation=approval_explanation,
-                tool_id=tool_id,
-                parent_tool_id=parent_tool_id,
-            )
-        )
-
-    async def on_credential_info(
-        self,
-        vault_path: str,
-        name: str,
-        detail: str,
-        approval_source: ApprovalSource | None = None,
-        approval_verdict: ApprovalVerdict | None = None,
-        approval_explanation: str | None = None,
-        tool_id: str | None = None,
-        parent_tool_id: str | None = None,
-    ) -> None:
-        await self._safe_send(
-            ToolCallInfo(
-                tool="credential_access",
-                args={"vault_path": vault_path, "name": name},
-                detail=detail,
-                approval_source=approval_source,
-                approval_verdict=approval_verdict,
-                approval_explanation=approval_explanation,
-                tool_id=tool_id,
-                parent_tool_id=parent_tool_id,
-            )
-        )
-
-    async def on_credential_approval_request(
-        self,
-        request_id: str,
-        vault_paths: list[str],
-        names: list[str],
-        descriptions: list[str],
-        skill_name: str | None,
-        explanation: str,
-    ) -> None:
-        await self._safe_send(
-            CredentialApprovalRequest(
-                request_id=request_id,
-                vault_paths=vault_paths,
-                names=names,
-                descriptions=descriptions,
-                skill_name=skill_name,
-                explanation=explanation,
-            )
-        )
-
-
-@router.websocket("/chat/{session_id}")
-async def chat_ws(
-    websocket: WebSocket,
-    session_id: str,
-    _token: Annotated[str, Depends(_verify_ws_token)],
-    client_id: Annotated[str | None, Query()] = None,
-) -> None:
-    current_state = _engine.session_mgr.load_state(session_id)
-    if current_state is None:
-        logger.warning(f"WebSocket rejected — session {session_id} not found")
-        await websocket.close(code=4004, reason="Session not found")
-        return
-
-    await websocket.accept()
-    logger.info(f"WebSocket connected for session {session_id}")
-
-    normalized_client_id = client_id.strip() if client_id else ""
-    ws_source_id = normalized_client_id or f"ws:{uuid.uuid4().hex}"
-    ws_client_type: NotificationClientType = "cli" if current_state.channel_type == "cli" else "web"
-    await _set_notification_presence(
-        session_id=session_id,
-        source_id=ws_source_id,
-        client_type=ws_client_type,
-        focus_state="visible",
-    )
-
-    sub = WebSocketSubscriber(websocket)
-    active = _engine.subscribe(session_id, sub)
-
-    # Tell the client whether an agent turn is in progress.
-    agent_running = active.agent_task is not None and not active.agent_task.done()
-    usage = _engine._turn_usage_payload(active)
-    with contextlib.suppress(Exception):
-        await _send(
-            websocket,
-            StatusUpdate(
-                agent_running=agent_running,
-                usage=usage,
-                llm_activity=_llm_activity_payload(active.llm_request_state if agent_running else None),
-            ),
-        )
-
-    # If agent is already running (e.g. reconnect), the subscriber will
-    # start receiving events immediately.  If there are pending approvals,
-    # re-send them so the client can respond.
-    for pa in list(active.pending_approval_requests):
-        with contextlib.suppress(Exception):
-            await _send(
-                websocket,
-                ApprovalRequest(
-                    tool_call_id=pa["tool_call_id"],
-                    tool=pa.get("tool", ""),
-                    args=pa.get("args", {}),
-                    explanation=pa.get("explanation", ""),
-                    risk_level=pa.get("risk_level", ""),
-                ),
-            )
-    for pp in list(active.pending_escalations):
-        with contextlib.suppress(Exception):
-            if pp.get("kind") == "git_push":
-                await _send(
-                    websocket,
-                    GitPushApprovalRequest(
-                        request_id=pp["request_id"],
-                        ref=pp.get("ref", ""),
-                        explanation=pp.get("explanation", ""),
-                        changed_files=pp.get("changed_files", []),
-                    ),
-                )
-            elif pp.get("kind") == "credential_access":
-                await _send(
-                    websocket,
-                    CredentialApprovalRequest(
-                        request_id=pp["request_id"],
-                        vault_paths=pp.get("vault_paths", []),
-                        names=pp.get("names", []),
-                        descriptions=pp.get("descriptions", []),
-                        skill_name=pp.get("skill_name"),
-                        explanation=pp.get("explanation", ""),
-                    ),
-                )
-            else:
-                await _send(
-                    websocket,
-                    DomainAccessApprovalRequest(
-                        request_id=pp["request_id"],
-                        domain=pp.get("domain", ""),
-                        command=pp.get("command", ""),
-                    ),
-                )
-
-    try:
-        while True:
-            raw = await websocket.receive_json()
-            try:
-                client_msg = parse_client_message(raw)
-            except (ValueError, Exception) as exc:
-                await _send(websocket, ErrorMessage(detail=str(exc)))
-                continue
-
-            current_state = _engine.session_mgr.load_state(session_id)
-            archived_session = current_state is not None and current_state.attributes.archived
-
-            # --- Cancel in-flight agent turn ---
-            if isinstance(client_msg, CancelRequest):
-                await _engine.submit_cancel(session_id)
-                continue
-
-            if archived_session and isinstance(client_msg, RetryLatestTurnRequest | ResetToTurnRequest | UserMessage):
-                await _send(websocket, ErrorMessage(detail="Archived sessions must be unarchived before use"))
-                continue
-
-            if isinstance(client_msg, RetryLatestTurnRequest):
-                await _engine.retry_latest_turn(session_id, origin=sub)
-                continue
-
-            if isinstance(client_msg, ResetToTurnRequest):
-                reset_applied = await _engine.reset_to_turn(session_id, client_msg.event_index)
-                if reset_applied:
-                    await _send(
-                        websocket,
-                        CommandResult(
-                            command="reset_to_turn",
-                            data={"event_index": client_msg.event_index},
-                        ),
-                    )
-                continue
-
-            # --- Approval responses — forward to engine ---
-            if isinstance(client_msg, ApprovalResponse | EscalationResponse):
-                await _engine.submit_approval(session_id, client_msg)
-                continue
-
-            if not isinstance(client_msg, UserMessage):
-                await _send(websocket, ErrorMessage(detail="Expected a message"))
-                continue
-
-            user_input = client_msg.content.strip()
-            if not user_input:
-                continue
-
-            # --- Slash commands ---
-            if user_input.startswith("/"):
-                if user_input.lower() in ("/quit", "/exit"):
-                    await websocket.close(code=1000)
-                    break
-
-                cmd_result = await _engine.handle_slash_command(session_id, user_input)
-                if cmd_result:
-                    result = CommandResult(
-                        command=cmd_result["command"],
-                        data=cmd_result["data"],
-                    )
-                    await _send(websocket, UserMessageNotification(content=user_input))
-                    await _send(websocket, result)
-                    _engine.session_mgr.append_events(
-                        session_id,
-                        [
-                            {"role": "user", "content": user_input},
-                            {"role": "command", "command": result.command, "data": result.data},
-                        ],
-                    )
-                    if result.command == "budget":
-                        await _send(
-                            websocket,
-                            StatusUpdate(
-                                agent_running=active.agent_task is not None and not active.agent_task.done(),
-                                usage=_engine._turn_usage_payload(active),
-                                llm_activity=_llm_activity_payload(active.llm_request_state),
-                            ),
-                        )
-                    continue
-
-            # --- Agent turn ---
-            await _engine.submit_message(session_id, user_input, origin=sub)
-
-    except WebSocketDisconnect as exc:
-        logger.info(f"Client disconnected from session {session_id} (code={exc.code})")
-    except Exception as exc:
-        logger.exception(f"Unexpected WebSocket error in session {session_id}: {exc}")
-        with contextlib.suppress(Exception):
-            await websocket.close(code=1011)
-    finally:
-        _notification_presence.remove_presence(session_id=session_id, source_id=ws_source_id)
-        _engine.unsubscribe(session_id, sub)
-        logger.debug(f"WebSocket cleanup for session {session_id}")
-
-
 class _InterceptHandler(logging.Handler):
     """Route stdlib logging records to loguru."""
 
@@ -2101,7 +1400,7 @@ async def evaluate_push(req: PushEvalRequest) -> dict[str, str]:
     if active.security is None or active.sentinel is None:
         return {"verdict": "deny", "reason": "Session not initialized"}
 
-    from .security import evaluate_push_with
+    from ..security import evaluate_push_with
 
     with _engine.llm_request_recording(active):
         try:
@@ -2125,6 +1424,8 @@ async def evaluate_push(req: PushEvalRequest) -> dict[str, str]:
     return {"verdict": "deny", "reason": "Denied by sentinel"}
 
 
+router.include_router(notifications_router)
+router.include_router(websocket_router)
 app.include_router(router)
 
 
@@ -2258,7 +1559,7 @@ async def fetch_credential(request: Request, vault_path: str) -> Response:
         if active.security is None or active.sentinel is None:
             return Response(status_code=403, content="Session not initialized")
 
-        from .security import evaluate_credential_with
+        from ..security import evaluate_credential_with
 
         with _engine.llm_request_recording(active):
             try:
